Involve expected library into project
This changeset modifies the `CMakeLists.txt` file in the
project. It adds a new setting `EXPECTED_BUILD_PACKAGE` and
`BUILD_TESTING` to `OFF`. Additionally, it includes the
`expected` library into the project by adding a new
subdirectory. The `target_link_libraries` now includes the
`expected` library along with other dependencies.