Add ProgramOptions class to CMake
A changeset has been made to the `CMakeLists.txt` file.
Specifically, a new class `ProgramOptions` has been added to
the project alongside the existing `error-handling` class.
This changeset includes the addition of `program_options.h`
and `program_options.cpp` files to the list of sources for
the `qt_add_qml_module` function.