Add ProgramOption class to parse command arguments
The changeset adds a new class called `ProgramOptions` to
parse command arguments in the `program_options.cpp` and
`program_options.h` files. In the `program_options.cpp`
file, a static method `parse_program_options` is defined
within the `gitse2` namespace. This method uses
`QCommandLineParser` to parse command line arguments and
sets up options for the utility. It also handles the case
where the `first_commit` argument is omitted. The
`ProgramOptions` class stores the `first_commit` argument as
a member variable.  In the `program_options.h` file, the
`ProgramOptions` class is declared with a private
constructor and a static method `parse_program_options`. The
class also includes a member variable `m_first_commit` to
store the parsed `first_commit` argument.