Parse program options in the beginning of the program
The changeset modifies the `main.cpp` file to include the
header file `program_options.h` and adds code to parse
program options at the beginning of the program. It calls
the `parse_program_options` function from the
`ProgramOptions` namespace to parse the options. If parsing
is unsuccessful, it logs an error message using
`qCritical()` and returns with an exit code of 1.