Introduce libgit2 initializer
This changeset introduces a new class `LibGit2Initializer`
in the `repository.cpp` file. This class is responsible for
initializing and shutting down the libgit2 library. It
includes methods to initialize the library when needed and
to shut it down when it's no longer in use. The class also
maintains a boolean flag to keep track of the initialization
status. Additionally, a global static instance of
`LibGit2Initializer` named `_initializer` is declared.  In
the `repository.h` file, a header guard is added to prevent
multiple inclusions of the file. It includes necessary
headers such as `<memory>`, `"error-handling.h"`,
`<git2.h>`, and `<QtCore/qglobal.h>` for the proper
functioning of the `LibGit2Initializer` class.