Open repository on program start
The changeset adds a new include directive for a header file
named "repository.h" in the main.cpp file. It then
initializes a repository object using the `open()` method
from the `Repository` class. If the repository object is not
successfully initialized, an error message is logged using
`qCritical()` and the program returns with an exit code of
2.