Introduce Repository class
This changeset introduces a new `Repository` class to the
project.   In the `CMakeLists.txt` file, the `repository.h`
and `repository.cpp` files are added to the list of sources
for the QML module.  In the `repository.cpp` file, the
`Repository` class is defined with a static method `open()`
that initializes the LibGit2 library, opens a repository at
a specified path, and handles errors related to repository
opening. The class also defines a destructor that frees the
repository when the object is destroyed.  In the
`repository.h` file, the `Repository` class is declared
within the `gitse2` namespace. It includes a shared pointer
type `RepositoryRef` for managing instances of the
`Repository` class. The class contains a static method
`open()` for opening a repository and a private constructor,
destructor, and member variables for managing the Git
repository.