Squash from first_commit to head to separate branch on program start
The changeset includes adding a new line of code to squash
changes from the `first_commit` to the head onto a separate
branch when the program starts. It introduces a new variable
`rval` that stores the result of squashing the changes. If
the squashing operation fails, an error message is logged,
and the program returns with an error code.