Always use [[nodiscard]]
The changeset modifies the `repository.h` file by adding the
`[[nodiscard]]` attribute to the `open()` method of the
`Repository` class. Additionally, a new line is added after
the `squash()` method declaration. This change ensures that
the compiler will produce a warning if the return value of
the `open()` method is unintentionally ignored, promoting
safer code practices.