Introduce SquashDiff: storage for diffs
This changeset introduces a new storage class called
`SquashDiff` by adding two new files: `squashdiff.cpp` and
`squashdiff.h`.   In `squashdiff.cpp`, the file is created
with the inclusion of `squashdiff.h` header file and the
usage of the `gitse2` namespace.  In `squashdiff.h`, a new
class `SquashDiff` is defined within the `gitse2` namespace.
This class contains a struct `diff_list_item` with various
data members related to storing diff information.
Additionally, a `diff_list` type alias is defined as a
vector of `diff_list_item`. The class also includes a
private member `m_diff_list` of type `diff_list`. The class
has a private constructor and is a friend of the
`Repository` class. Finally, a type alias `SquashDiffPtr` is
defined as a unique pointer to `SquashDiff`.  These changes
set up the foundation for storing and managing diff
information within the `SquashDiff` class.