Introduce a method that creates SquashDiff
A `changeset` has been introduced to the `repository.cpp`
file which adds a new method `create_squash_diff()` to the
`Repository` class. This method is responsible for creating
a `SquashDiff` object by performing various Git operations
such as obtaining tree pointers, calculating the diff
between two trees, and populating the `SquashDiff` object
with relevant data. Additionally, error handling and
callbacks for processing file and binary differences are
implemented within this method.  In the `repository.h` file,
a `changeset` has been made to include the header file
`"squashdiff.h"` in the `Repository` class. This change
allows the declaration of the new method
`create_squash_diff()` which returns a `Result` containing a
pointer to a `SquashDiff` object.