Repository remembers first commit now
The changeset includes modifications in both
`repository.cpp` and `repository.h` files. In
`repository.cpp`, the `new_head` variable is replaced with
`m_first_commit` to store the first commit in the
repository. The `new_branch_name` and `create_branch`
function now use `m_first_commit` instead of `new_head`.
Additionally, the `commit2` variable is updated to use
`m_first_commit` instead of `new_head` in the `squash`
function.  In `repository.h`, a new member variable
`m_first_commit` of type `git_commit_ptr` is added to the
`Repository` class to store the first commit in the
repository.