Method `apply_diff` is now using GIT_DIFF_SHOW_BINARY flag
The changeset modifies the `apply_diff` method in the
`repository.cpp` file. It introduces a new
`git_diff_options` variable `opts` initialized with
`GIT_DIFF_OPTIONS_INIT`. The `GIT_DIFF_SHOW_BINARY` flag is
set in the `opts` variable. This flag is then passed to the
`git_diff_tree_to_tree` function as a parameter to indicate
that binary files should be shown in the diff.