Fix path to Main.qml
The changeset updates the path to the Main.qml file in the main.cpp file. The
previous path was "qrc:/git-se2/Main.qml", and it has been changed to
"qrc:/ldd/so/gitse2/Main.qml". This modification ensures that the correct
Main.qml file is referenced and loaded in the application.