Rename project name to get rid of hyphen
The changeset modifies the project name in the CMakeLists.txt file to remove the
hyphen. It updates occurrences of `git-se2` to `gitse2` throughout the file,
including changing the project name, executable name, QML module URI, target
properties, include directories, and target link libraries. Additionally, it
adjusts the QML module URI to `ldd.so.gitse2` and updates the bundle version and
short version string properties for macOS.