Repository is now using new `DiffListItem` class
The changeset introduces the usage of the new `DiffListItem` class in the
repository. It includes modifications in the `repository.cpp` file where the
`SquashDiff::diff_list` is replaced with `DiffList`, and
`SquashDiff::diff_list_item` is replaced with `DiffListItemPtr`. The `add_delta`
and `add_binary_delta` methods are used to add delta and binary delta
information to the `DiffListItem` objects. Additionally, some unnecessary code
related to paths and binary data handling has been removed.