SquashDiff becomes QObject
It is also exposing a new property `title`

This changeset modifies the `SquashDiff` class in both the `squashdiff.cpp` and
`squashdiff.h` files. In `squashdiff.h`, the class `SquashDiff` now inherits
from `QObject` and includes the necessary headers for `QString` and `QObject`. A
new property `title` is exposed using the `Q_PROPERTY` macro, with a
corresponding getter function `diff_title()` and a signal `diff_title_changed`.
Additionally, a member variable `m_diff_title` of type `QString` is added with a
default value of "Untitled". The getter function `diff_title()` returns a
constant reference to `m_diff_title`.