Connect tree model to QML
This changeset connects the tree model to QML by adding a new method
`tree_model()` in the `SquashDiff` class in `squashdiff.cpp` file. This method
returns the `m_tree_model` object. Additionally, a new signal
`tree_model_changed` is added to the class declaration in the `squashdiff.h`
file to notify when the tree model changes. The `m_tree_model` member variable
is also added to the `SquashDiff` class. The `initialize()` method now creates
the model using the `m_tree_model` and returns an error if the model creation
fails.