Repository initializes SquashDiff
Repository is a factory for SquashDiff. It is decided that it will call
`initialize` method as a constructor

The changeset includes modifications in three files.

In `repository.cpp`, a new condition is added to call the `initialize` method of
`SquashDiff` before returning the `SquashDiffPtr` object. If the initialization
fails, an error is returned.

In `squashdiff.cpp`, the `initialize` method is defined. Inside this method, the
`m_diff_list` is sorted based on the count of slashes in the path of each
element. After sorting, the `m_tree_model` is assigned a value, and an empty
`Result` is returned.

In `squashdiff.h`, the `initialize` method is declared as `[[nodiscard]]`,
indicating that the method should not be ignored by the caller. Additionally, an
include statement for `error-handling.h` is added to import necessary error
handling functionalities.