Introduce `DiffTreeItem` class
`DiffTreeItem` class is a tree item for the Qt treeview qml

This changeset introduces a new class called `DiffTreeItem`. This class is
designed to be used as a tree item in the Qt treeview QML. The changeset
includes the addition of `difftreeitem.cpp` and `difftreeitem.h` files to the
project.